task-060 Updating entities

+++ src/main/java/com/app/diamondhotelbackend/entity/Transaction.java
@@ -33,7 +33,7 @@ public class Transaction {
 
     private BigDecimal cost;
 
-    private BigDecimal tax;
+    private double tax;
 
     private String status;
